+++ examples/live-chat/src/live-demo/TUILive/index.ts
@@ -1,6 +1,0 @@
-export * from './TUILive';
-export * from './TUILiveHeader';
-export * from './TUILiveContent';
-export * from './TUILiveFooter';
-export * from './TUILiveMemberList';
-export * from './context/TUILiveContext';