+++ src/locales/en-US/TUIGlobal.ts
@@ -1,6 +1,0 @@
-/* eslint-disable quote-props */
-const TUIGlobal = {
-  'chats': 'Messages',
-  'contacts': 'Contacts',
-}
-export default TUIGlobal;