+++ src/locales/ja-JP/TUIGlobal.ts
@@ -1,6 +1,0 @@
-/* eslint-disable quote-props */
-export const TUIGlobal = {
-  'chats': 'メッセージ',
-  'contacts': '連絡先',
-}
-export default TUIGlobal;