+++ src/locales/ja-JP/index.ts
@@ -1,11 +1,9 @@
-import TUIGlobal from './TUIGlobal';
 import TUIChat from './TUIChat';
 import TUIConversation from './TUIConversation';
 import TUIContact from './TUIContact';
 import TUIProfile from './TUIProfile';
 
 export const message = {
-  TUIGlobal,
   TUIChat,
   TUIConversation,
   TUIContact,