+++ src/locales/ko-KR/TUIGlobal.ts
@@ -1,6 +1,0 @@
-/* eslint-disable quote-props */
-export const TUIGlobal = {
-  'chats': '채팅',
-  'contacts': '연락처'
-}
-export default TUIGlobal;